+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqlCompiler/ir/expression/literal/DBSPU64Literal.java
@@ -34,8 +34,6 @@ import org.dbsp.util.IIndentStream;
 import javax.annotation.Nullable;
 import java.util.Objects;
 
-import static org.dbsp.sqlCompiler.ir.type.DBSPTypeCode.UINT64;
-
 public class DBSPU64Literal extends DBSPIntLiteral {
     @Nullable
     public final Long value;
@@ -58,7 +56,7 @@ public class DBSPU64Literal extends DBSPIntLiteral {
     }
 
     public DBSPU64Literal(CalciteObject node, @Nullable Long value, boolean nullable) {
-        this(node, new DBSPTypeInteger(CalciteObject.EMPTY, UINT64,64, false, nullable), value);
+        this(node, new DBSPTypeInteger(CalciteObject.EMPTY, 64, false, nullable), value);
         if (value == null && !nullable)
             throw new InternalCompilerError("Null value with non-nullable type", this);
     }